pyproject.toml
[tool.poetry]
name = "academic-ai-backend"
version = "0.1.0"
description = ""
authors = ["Ajju-kld <mohammedajmal996@gmail.com>"]
readme = "README.md"
packages = [{include = "academic_ai_backend"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.15.0"
django-cors-headers = "^4.3.1"
django-rest-knox = "^4.2.0"
torch = "^2.2.1"
numpy = "^1.26.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
